Fixed game crashing bug
- Navigating to the game page now actually calls the `/game/gameStart`
backend route
- Shuffled frontend routes around to new files in `services`
- Edited `components/InputBox.jsx` so that it uses the correct frontend
service
- `pages/game/Game.jsx` now actually communicates with the backend when
starting a new game, submitting guesses etc.
